app/app/requirements.txt
Flask==2.2.2
psycopg2-binary==2.9.5
numpy==1.23.4
